refactor: 컨트롤러로 분리했던 게임의 흐름 도메인으로 이동
