+++ src/main/java/Main.java
@@ -1,8 +1,15 @@
-import config.Config;
+import domain.RacingGame;
+import view.InputView;
+import view.MessageResolver;
+import view.OutputView;
 
 public class Main {
     public static void main(String[] args) {
-        Config config = new Config();
-        config.racingGameController().run();
+        InputView inputView = new InputView();
+        MessageResolver messageResolver = new MessageResolver();
+        OutputView outputView = new OutputView(messageResolver);
+        RacingGame racingGame = new RacingGame(inputView, outputView);
+
+        racingGame.run();
     }
 }
